refactor: modularize miku runtime and sqlite indexing
