+++ crates/miku-index-sqlite/Cargo.toml
@@ -12,8 +12,7 @@ miku-domain = { version = "0.0.2", path = "../miku-domain" }
 miku-indexer = { version = "0.0.2", path = "../miku-indexer" }
 serde_json.workspace = true
 sqlx = { version = "0.8.6", default-features = false, features = ["sqlite", "runtime-tokio", "migrate", "macros"] }
-tokio = { workspace = true, features = ["macros"] }
 
 [dev-dependencies]
 tempfile = "3"
-tokio = { workspace = true, features = ["rt-multi-thread"] }
+tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
